+++ packages/ckb-sdk-rpc/src/resultFormatter.ts
@@ -1,27 +1,27 @@
 /* eslint-disable camelcase */
-const formatters = {
-  toNumber: (number: string | number) => +number,
-  toHash: (hash: string) => hash,
+const formatter = {
+  toNumber: (number: CKB_RPC.BlockNumber): CKBComponents.BlockNumber => number,
+  toHash: (hash: CKB_RPC.Hash256): CKBComponents.Hash256 => hash,
   toHeader: ({
     version,
     parent_hash,
     timestamp,
     number,
-    txs_commit,
-    txs_proposal,
+    transactions_root,
+    proposals_root,
     witnesses_root,
     difficulty,
     uncles_hash,
     uncles_count,
     seal,
     hash,
-  }: any) => ({
+  }: CKB_RPC.Header): CKBComponents.BlockHeader => ({
     version,
     number,
     parentHash: parent_hash,
     timestamp,
-    txsCommit: txs_commit,
-    txsProposal: txs_proposal,
+    transactionsRoot: transactions_root,
+    proposalsRoot: proposals_root,
     witnessesRoot: witnesses_root,
     difficulty,
     unclesHash: uncles_hash,
@@ -29,50 +29,78 @@ const formatters = {
     seal,
     hash,
   }),
-  toTransactionWithHash: (txn: any) => txn,
-  toCells: (
-    cells: {
-      capacity: number
-      lock: string
-      out_point: {
-        hash: string
-        index: number
-      }
-    }[]
-  ) =>
-    cells.map(({ capacity, lock, out_point }) => ({
-      capacity,
-      lock,
-      outPoint: out_point,
-    })),
-  toCellOutputWithOutput: (cell: any) => cell,
-  toCellWithStatus: (cell: any) => cell,
-  toBlockWithHash: (block: any) => block,
-  toTxRes: (txRes: any) => txRes,
-  toBlock: ({ header, uncles, commit_transactions, proposal_transactions }: any) => {
-    const h: CKBComponents.BlockHeader = formatters.toHeader(header)
-    const u = uncles
-    const c = commit_transactions
-    const p = proposal_transactions
-    const b: CKBComponents.Block = {
-      header: h,
-      uncles: u,
-      commitTransactions: c,
-      proposalTransactions: p,
-    }
-    return b
-  },
-  toTrace: (trace: CKBComponents.TransactionTrace) => trace,
-  toLocalNodeInfo: (info: {
-    version: string
-    node_id: string
-    addresses: { address: string; score: number }[]
-  }): CKBComponents.NodeInfo => ({
-    version: info.version,
-    nodeId: info.node_id,
-    addresses: info.addresses,
+  toScript: ({ args, code_hash: codeHash }: CKB_RPC.Script): CKBComponents.Script => ({
+    args,
+    codeHash,
+  }),
+  toInput: ({ previous_output: previousOutput, since, args }: CKB_RPC.CellInput): CKBComponents.CellInput => ({
+    previousOutput: formatter.toOutPoint(previousOutput),
+    since,
+    args,
+  }),
+  toOutput: ({ capacity, data, lock, type }: CKB_RPC.CellOutput): CKBComponents.CellOutput => ({
+    capacity,
+    data,
+    lock: formatter.toScript(lock),
+    type: type ? formatter.toScript(type) : null,
+  }),
+  toOutPoint: ({ tx_hash: txHash, index }: CKB_RPC.OutPoint): CKBComponents.OutPoint => ({
+    txHash,
+    index,
+  }),
+  toTransaction: ({
+    deps,
+    inputs,
+    outputs,
+    version,
+    witnesses,
+    hash,
+  }: CKB_RPC.Transaction): CKBComponents.Transaction => ({
+    deps: deps.map(formatter.toOutPoint),
+    inputs: inputs.map(formatter.toInput),
+    outputs: outputs.map(formatter.toOutput),
+    version,
+    witnesses,
+    hash,
+  }),
+  toUncleBlock: ({ header, proposals }: CKB_RPC.UncleBlock): CKBComponents.UncleBlock => ({
+    header: formatter.toHeader(header),
+    proposals,
   }),
 
-  /* eslint-enable camelcase */
+  toBlock: ({ header, uncles, transactions, proposals }: CKB_RPC.Block): CKBComponents.Block => ({
+    header: formatter.toHeader(header),
+    uncles: uncles.map(formatter.toUncleBlock),
+    transactions: transactions.map(formatter.toTransaction),
+    proposals,
+  }),
+  toTrace: (trace: CKBComponents.TransactionTrace) => trace,
+  toNodeInfo: ({ addresses, node_id: nodeId, version }: CKB_RPC.NodeInfo): CKBComponents.NodeInfo => ({
+    addresses,
+    nodeId,
+    version,
+  }),
+  toCell: ({ capacity, data, lock, type }: CKB_RPC.Cell): CKBComponents.Cell => ({
+    capacity,
+    data,
+    lock: formatter.toScript(lock),
+    type: type ? formatter.toScript(type) : null,
+  }),
+  toCellWithStatus: ({ cell, status }: { cell: CKB_RPC.Cell; status: string }) => ({
+    cell: formatter.toCell(cell),
+    status,
+  }),
+  toCells: (cells: CKB_RPC.Cell[]): CKBComponents.Cell[] => cells.map(formatter.toCell),
+  toTransactionWithStatus: ({
+    transaction,
+    tx_status: { block_hash: blockHash, status },
+  }: CKB_RPC.TransactionWithStatus) => ({
+    transaction: formatter.toTransaction(transaction),
+    txStatus: {
+      blockHash,
+      status,
+    },
+  }),
 }
-export default formatters
+export default formatter
+/* eslint-enable camelcase */